moved filter function outside struct
