+++ crates/client/rpc/src/lib.rs
@@ -7,8 +7,13 @@ mod errors;
 mod madara_backend_client;
 mod types;
 
+#[cfg(test)]
+mod tests;
+
+use std::iter::Skip;
 use std::marker::PhantomData;
 use std::sync::Arc;
+use std::vec::IntoIter;
 
 use errors::StarknetRpcApiError;
 use jsonrpsee::core::{async_trait, RpcResult};
@@ -185,52 +190,6 @@ where
         Ok((block_events, block))
     }
 
-    /// Helper function to get filter events using address and keys
-    ///
-    /// # Arguments
-    ///
-    /// * `events` - A vector of all events
-    /// * `address` - Address to use to filter the events
-    /// * `keys` - Keys to use to filter the events. An event is filtered if any key is present
-    /// * `max_results` - Optional, indicated the max events that need to be filtered
-    ///
-    /// # Returns
-    ///
-    /// * `(block_events: Vec<EventWrapper>, continuation_token: usize)` - A tuple of the filtered
-    ///   events and the first index which still hasn't been processed block_id and an instance of
-    ///   Block
-    pub fn filter_events_by_params(
-        &self,
-        events: Vec<EventWrapper>,
-        address: Option<Felt252Wrapper>,
-        keys: Vec<Vec<FieldElement>>,
-        max_results: Option<usize>,
-    ) -> (Vec<EventWrapper>, usize) {
-        let mut filtered_events = vec![];
-        let mut index = 0;
-
-        // Iterate on block events.
-        for event in events {
-            index += 1;
-            let match_from_address = address.map_or(true, |add| add == event.from_address);
-            // Based on https://github.com/starkware-libs/papyrus
-            let match_keys = keys
-                .iter()
-                .enumerate()
-                .all(|(i, keys)| event.keys.len() > i && (keys.is_empty() || keys.contains(&event.keys[i].into())));
-
-            if match_from_address && match_keys {
-                filtered_events.push(event);
-                if let Some(max_results) = max_results {
-                    if filtered_events.len() >= max_results {
-                        break;
-                    }
-                }
-            }
-        }
-        (filtered_events, index)
-    }
-
     /// Helper function to filter Starknet evetomnts provided a RPC event filter
     ///
     /// # Arguments
@@ -264,7 +223,7 @@ where
                 continue;
             }
 
-            let block_events = block_events.into_iter().skip(continuation_token).collect();
+            let block_events = block_events.into_iter().skip(continuation_token);
             // Kept in order to calculate continuation token.
             let block_events_len = block_events_len - continuation_token;
             let index_before_loop = index;
@@ -275,7 +234,7 @@ where
                 StarknetRpcApiError::BlockNotFound
             })?;
 
-            let (new_filtered_events, continuation_index) = self.filter_events_by_params(
+            let (new_filtered_events, continuation_index) = filter_events_by_params(
                 block_events,
                 from_address,
                 keys.clone(),
@@ -1002,4 +961,48 @@ where
             None => Err(StarknetRpcApiError::TxnHashNotFound.into()),
         }
     }
+}
+
+/// Helper function to get filter events using address and keys
+///
+/// # Arguments
+///
+/// * `events` - A vector of all events
+/// * `address` - Address to use to filter the events
+/// * `keys` - Keys to use to filter the events. An event is filtered if any key is present
+/// * `max_results` - Optional, indicated the max events that need to be filtered
+///
+/// # Returns
+///
+/// * `(block_events: Vec<EventWrapper>, continuation_token: usize)` - A tuple of the filtered
+///   events and the first index which still hasn't been processed block_id and an instance of Block
+pub fn filter_events_by_params(
+    events: Skip<IntoIter<EventWrapper>>,
+    address: Option<Felt252Wrapper>,
+    keys: Vec<Vec<FieldElement>>,
+    max_results: Option<usize>,
+) -> (Vec<EventWrapper>, usize) {
+    let mut filtered_events = vec![];
+    let mut index = 0;
+
+    // Iterate on block events.
+    for event in events {
+        index += 1;
+        let match_from_address = address.map_or(true, |add| add == event.from_address);
+        // Based on https://github.com/starkware-libs/papyrus
+        let match_keys = keys
+            .iter()
+            .enumerate()
+            .all(|(i, keys)| event.keys.len() > i && (keys.is_empty() || keys.contains(&event.keys[i].into())));
+
+        if match_from_address && match_keys {
+            filtered_events.push(event);
+            if let Some(max_results) = max_results {
+                if filtered_events.len() >= max_results {
+                    break;
+                }
+            }
+        }
+    }
+    (filtered_events, index)
 }
