+++ crates/node/src/main.rs
@@ -10,8 +10,6 @@ mod command;
 mod constants;
 mod rpc;
 mod starknet;
-#[cfg(test)]
-mod tests;
 
 fn main() -> sc_cli::Result<()> {
     command::run()
